Minor fixes

+++ src/main/java/com/appslandia/common/jdbc/SqlTypeMapper.java
@@ -52,7 +52,7 @@ public class SqlTypeMapper {
 	map.put(Types.BIGINT, long.class);
 
 	map.put(Types.REAL, float.class);
-	map.put(Types.FLOAT, double.class);
+	map.put(Types.FLOAT, float.class);
 	map.put(Types.DOUBLE, double.class);
 	map.put(Types.NUMERIC, BigDecimal.class);
 	map.put(Types.DECIMAL, BigDecimal.class);
